refactor(eventCheckout): separate event checkout logic into its own module and update references in the service
